fix: sidenav mobile test
